requirements.txt
flask
Flask-SQLAlchemy
Flask-Security-Too
email-validator
argon2_cffi
dotenv
requests
shortuuid
